Remove website-beta directory and migrate files
Deleted all files in the website-beta directory and moved relevant scripts and source files to the main website directory. This change consolidates code and assets, eliminating duplication and streamlining project structure.
