--- website-beta/.prettierignore
@@ -1,4 +1,0 @@
-.prettierrc
-node_modules
-pnpm-lock.yaml
-._*
