--- website-beta/README.md
@@ -1,7 +1,0 @@
-# Website (Cloudflare Worker + Hono + React)
-
-## 環境変数
-
-```env
-SECRET_TOKEN="YOUR_TOKEN"
-```
