--- website-beta/attribution.md
@@ -1,20 +1,0 @@
-# Open Source Attribution for Scratch Status Monitor
-
-- Hono
-- ky
-- react
-- react dom
-- zod
-
-- @cloudflare/vite-plugin
-- @cloudflare/workers-types
-- @tailwindcss/postcss
-- @types/react
-- @types/react-dom
-- @vitejs/plugin-react
-- postcss
-- tailwindcss
-- vite
-- vite-ssr-components
-- vite-tsconfig-paths
-- wrangler
