--- website-beta/package.json
@@ -1,27 +1,0 @@
-{
-  "name": "website",
-  "type": "module",
-  "scripts": {
-    "dev": "wrangler dev",
-    "deploy": "wrangler deploy --minify",
-    "cf-typegen": "wrangler types --env-interface CloudflareBindings",
-    "prettier": "pnpm dlx prettier . --write"
-  },
-  "dependencies": {
-    "hono": "^4.10.6",
-    "ky": "^1.14.0",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "zod": "^4.1.13"
-  },
-  "devDependencies": {
-    "@cloudflare/vite-plugin": "^1.2.3",
-    "@cloudflare/workers-types": "^4.20251126.0",
-    "@tailwindcss/postcss": "^4.1.17",
-    "@types/react": "^19.2.7",
-    "@types/react-dom": "^19.2.3",
-    "postcss": "^8.5.6",
-    "tailwindcss": "^4.1.17",
-    "wrangler": "^4.17.0"
-  }
-}
