--- website-beta/project.config.ts
@@ -1,3 +1,0 @@
-export const projectConfig = {
-  origin: "scratch-status-monitor.scratchcore.org",
-};
