--- website-beta/public/.assetsignore
@@ -1,1 +1,0 @@
-.vite
