--- website-beta/src/index.tsx
@@ -1,124 +1,0 @@
-import { projectConfig } from "~/project.config";
-
-import { Hono } from "hono";
-import { prettyJSON } from "hono/pretty-json";
-import { trimTrailingSlash } from "hono/trailing-slash";
-import { languageDetector } from "hono/language";
-import { cors } from "hono/cors";
-import { csrf } from "hono/csrf";
-import { HTTPException } from "hono/http-exception";
-
-import { ZodError } from "zod";
-import { createNotFoundError } from "./schemas/NotFoundError";
-import { createUnprocessableEntityError } from "./schemas/UnprocessableEntity";
-import { createBadRequestError } from "./schemas/BadRequestError";
-
-import { MainRoutes } from "./routes/route";
-
-const app = new Hono();
-
-app.use(prettyJSON());
-app.use(trimTrailingSlash());
-
-// ユーザー言語: (https://hono.dev/docs/middleware/builtin/language)
-app.use(
-  languageDetector({
-    supportedLanguages: ["ja", "en"], // Must include fallback
-    fallbackLanguage: "ja", // Required
-  }),
-);
-
-// CORS: (https://hono.dev/docs/middleware/builtin/cors)
-app.use(
-  "/*",
-  cors({
-    origin: `https://${projectConfig.origin}`,
-    allowHeaders: ["X-Custom-Header", "Upgrade-Insecure-Requests"],
-    allowMethods: ["POST", "GET", "OPTIONS"],
-    exposeHeaders: ["Content-Length", "X-Kuma-Revision"],
-    maxAge: 600,
-    credentials: true,
-  }),
-);
-
-// CSRF保護 (https://hono.dev/docs/middleware/builtin/csrf)
-app.use(
-  csrf({
-    origin: projectConfig.origin,
-  }),
-);
-
-// 例外処理: (https://hono.dev/docs/api/exception)
-app.notFound((c) => c.json(createNotFoundError(), 404));
-
-app.onError((err, c) => {
-  if (err instanceof HTTPException) {
-    return err.getResponse();
-  }
-  if (err instanceof ZodError) {
-    return c.json(
-      createUnprocessableEntityError({
-        title: "バリデーションエラーが発生しました。",
-        errors: err.issues.map((issue) => ({
-          path: issue.path.join("."),
-          message: issue.message,
-        })),
-      }),
-      422,
-    );
-  }
-  return c.json(
-    createBadRequestError({
-      title: "Internal Server Error",
-      detail: err instanceof Error ? err.message : String(err),
-    }),
-    500,
-  );
-});
-
-app.route("/", MainRoutes);
-
-// Cloudflare Cron Triggers scheduled handler
-async function scheduled(_controller: any, env?: any, _ctx?: any) {
-  try {
-    const { checkStatus } = await import("~/src/routes/v1/api/status");
-    const { getStatusCache, setStatusCache } = await import(
-      "~/src/lib/statusCache"
-    );
-    const cache = await getStatusCache(env);
-    const cacheMinutes = cache.cacheMinutes || 3;
-    // Determine cron interval minutes once: prefer previously detected value,
-    // otherwise estimate from the previous cache timestamp if available.
-    let cronIntervalMinutes: number | null = cache.cronIntervalMinutes ?? null;
-    if (!cronIntervalMinutes && cache.ts && cache.ts > 0) {
-      const estimated = Math.max(1, Math.round((Date.now() - cache.ts) / 60000));
-      // Accept estimation only if reasonable (<= 60 minutes)
-      if (estimated > 0 && estimated <= 60) cronIntervalMinutes = estimated;
-    }
-
-    const monitors = await checkStatus();
-    // Compute explicit nextGenTs aligned to the detected/estimated interval so
-    // clients can rely on an exact timestamp for their countdowns.
-    let nextGenTs: number | null = null;
-    const now = Date.now();
-    const stepMinutes = cronIntervalMinutes ?? cacheMinutes;
-    if (stepMinutes && stepMinutes > 0) {
-      const minuteIndex = Math.floor(now / 60000);
-      const nextMultiple = Math.floor(minuteIndex / stepMinutes) * stepMinutes + stepMinutes;
-      nextGenTs = nextMultiple * 60000;
-      if (nextGenTs <= now) nextGenTs += stepMinutes * 60 * 1000;
-    }
-
-    // Pass both nextGenTs and cronIntervalMinutes so KV stores exact timestamp
-    // and the detected interval for future alignment.
-    await setStatusCache(monitors, Date.now(), cacheMinutes, env, nextGenTs, cronIntervalMinutes);
-    console.log("scheduled: refreshed status cache");
-  } catch (e) {
-    console.error("scheduled handler failed", e);
-  }
-}
-
-export default {
-  fetch: app.fetch,
-  scheduled,
-};
