--- website-beta/src/routes/v1/api/route.ts
@@ -1,6 +1,0 @@
-import { Hono } from "hono";
-import { statusRoute } from "./status";
-
-export const apiRoutes = new Hono();
-
-apiRoutes.route("/status", statusRoute);
