--- website-beta/src/routes/v1/route.ts
@@ -1,6 +1,0 @@
-import { Hono } from "hono";
-import { apiRoutes } from "./api/route";
-
-export const v1Routes = new Hono();
-
-v1Routes.route("/api", apiRoutes);
