--- website-beta/wrangler.jsonc
@@ -1,32 +1,0 @@
-{
-  "$schema": "node_modules/wrangler/config-schema.json",
-  "name": "scratch-status-monitor",
-  "compatibility_date": "2025-08-03",
-  "main": "./src/index.tsx",
-  "assets": {
-    "directory": "public",
-  },
-  // "triggers": {
-  "triggers": {
-    "crons": ["*/3 * * * *"],
-  },
-  "kv_namespaces": [
-    {
-      "binding": "SCRATCH_STATUS_MONITOR",
-      "id": "ad22b2fa63544c838e94dfe4d27634e8",
-    },
-  ],
-  "observability": {
-    "logs": {
-      "enabled": true,
-      "head_sampling_rate": 1,
-      "invocation_logs": true,
-      "persist": true,
-    },
-    "traces": {
-      "enabled": false,
-      "persist": true,
-      "head_sampling_rate": 1,
-    },
-  },
-}
