+++ website/src/routes/root.ts
@@ -1,52 +1,49 @@
 import { Hono } from "hono";
 import { RenderPage } from "../ssr/render";
-import { checkStatus } from "./v1/api/status";
+import { getStatusCache } from "~/src/lib/statusCache";
 
 export const RootPage = new Hono();
 
-let cached: { ts: number; data: any[] | null } = { ts: 0, data: null };
 // Cache duration in minutes
-let cacheMinutes = 1;
+let cacheMinutes = 3; // Use 3 minutes as the initial/fallback cacheMinutes in root SSR to match final-fallback policy
 
 // Root: server-rendered page using React SSR helper
 RootPage.get("/", async (c) => {
-  const force = c.req.query("refresh") === "1" || c.req.query("force") === "1";
-  const now = Date.now();
-
-  const age = cached.ts ? Math.floor((now - cached.ts) / 1000) : 0;
+  const cached = await getStatusCache(c.env);
   let results: any[] = [];
   let nextGenTs: number | null = null;
 
-  // Simple caching: treat cacheMinutes as the TTL in minutes.
-  if (!force && cached.data) {
-    if (age <= cacheMinutes * 60) {
-      results = cached.data as any[];
-    } else {
-      results = await checkStatus();
-      cached = { ts: Date.now(), data: results };
-    }
+  // Use server-side cache if available. Cron Triggers refresh the cache.
+  if (cached.monitors) {
+    results = cached.monitors as any[];
   } else {
-    results = await checkStatus();
-    cached = { ts: Date.now(), data: results };
+    // Cache not yet populated by Cron; render an empty state.
+    results = [];
   }
 
   // next generation timestamp = last cached timestamp + TTL (in ms)
-  if (cached.ts > 0) {
+  const updatedCache = await getStatusCache(c.env);
+  // prefer an explicitly stored nextGenTs first (scheduled handler writes it),
+  // then prefer detected cronIntervalMinutes or cacheMinutes for alignment.
+  cacheMinutes = updatedCache.cronIntervalMinutes ?? updatedCache.cacheMinutes ?? cacheMinutes;
+  if (updatedCache.nextGenTs && updatedCache.nextGenTs > updatedCache.ts) {
+    nextGenTs = updatedCache.nextGenTs;
+  } else if (updatedCache.ts > 0) {
     // Align nextGenTs to the next minute boundary that is a multiple of cacheMinutes.
     // Example: cacheMinutes=1 -> next minute (12:01:00). cacheMinutes=2 -> next even 2-minute boundary (12:02:00, 12:04:00, ...).
-    const minuteIndex = Math.floor(cached.ts / 60000); // minutes since epoch
+    const minuteIndex = Math.floor(updatedCache.ts / 60000); // minutes since epoch
     const nextMultiple =
       Math.floor(minuteIndex / cacheMinutes) * cacheMinutes + cacheMinutes;
     nextGenTs = nextMultiple * 60000;
     // Ensure nextGenTs is strictly in the future relative to cached.ts
-    if (nextGenTs <= cached.ts) {
+    if (nextGenTs <= updatedCache.ts) {
       nextGenTs += cacheMinutes * 60 * 1000;
     }
   }
 
   const html = RenderPage({
     monitors: results,
-    lastUpdated: cached.ts,
+    lastUpdated: updatedCache.ts,
     nextGenTs,
     cacheMinutes,
   });
